+++ src/main/java/fr/insee/survey/datacollectionmanagement/contact/controller/ContactEventController.java
@@ -1,7 +1,7 @@
 package fr.insee.survey.datacollectionmanagement.contact.controller;
 
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -66,12 +66,15 @@ public class ContactEventController {
     })
     public ResponseEntity<?> getContactContactEvents(@PathVariable("id") String identifier) {
         try {
-            Contact contact = contactService.findByIdentifier(identifier);
-            return ResponseEntity.status(HttpStatus.OK)
-                    .body(contact.getContactEvents().stream().map(ce -> convertToDto(ce)).collect(Collectors.toList()));
 
-        } catch (NoSuchElementException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Contact does not exist");
+            Optional<Contact> optContact = contactService.findByIdentifier(identifier);
+            if (optContact.isPresent()) {
+                return ResponseEntity.status(HttpStatus.OK)
+                        .body(optContact.get().getContactEvents().stream().map(ce -> convertToDto(ce))
+                                .collect(Collectors.toList()));
+
+            } else
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Contact does not exist");
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Error");
         }
@@ -87,25 +90,27 @@ public class ContactEventController {
     })
     public ResponseEntity<?> postContactEvent(@RequestBody ContactEventDto contactEventDto) {
         try {
-            Contact contact = contactService.findByIdentifier(contactEventDto.getIdentifier());
-            ContactEvent contactEvent = convertToEntity(contactEventDto);
-            ContactEvent newContactEvent = contactEventService.saveContactEvent(contactEvent);
-            Set<ContactEvent> setContactEvents = contact.getContactEvents();
-            setContactEvents.add(newContactEvent);
-            contact.setContactEvents(setContactEvents);
-            contactService.saveContact(contact);
-            HttpHeaders responseHeaders = new HttpHeaders();
-            responseHeaders.set(HttpHeaders.LOCATION, ServletUriComponentsBuilder.fromCurrentRequest().toUriString());
-            return ResponseEntity.status(HttpStatus.CREATED).headers(responseHeaders)
-                    .body(convertToDto(newContactEvent));
-        } catch (NoSuchElementException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Contact does not exist");
 
+            Optional<Contact> optContact = contactService.findByIdentifier(contactEventDto.getIdentifier());
+            if (optContact.isPresent()) {
+                Contact contact = optContact.get();
+                ContactEvent contactEvent = convertToEntity(contactEventDto);
+                ContactEvent newContactEvent = contactEventService.saveContactEvent(contactEvent);
+                Set<ContactEvent> setContactEvents = contact.getContactEvents();
+                setContactEvents.add(newContactEvent);
+                contact.setContactEvents(setContactEvents);
+                contactService.saveContact(contact);
+                HttpHeaders responseHeaders = new HttpHeaders();
+                responseHeaders.set(HttpHeaders.LOCATION,
+                        ServletUriComponentsBuilder.fromCurrentRequest().toUriString());
+                return ResponseEntity.status(HttpStatus.CREATED).headers(responseHeaders)
+                        .body(convertToDto(newContactEvent));
+            } else
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Contact does not exist");
         } catch (EventException e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body("Event not recognized: only [" + Stream.of(ContactEventType.values()).
-                            map(ContactEventType::name).
-                            collect(Collectors.joining(", "))+"] are possible");
+                    .body("Event not recognized: only [" + Stream.of(ContactEventType.values())
+                            .map(ContactEventType::name).collect(Collectors.joining(", ")) + "] are possible");
         }
 
     }
@@ -119,15 +124,17 @@ public class ContactEventController {
     })
     public ResponseEntity<?> deleteContactEvent(@PathVariable("id") Long id) {
         try {
-            ContactEvent contactEvent = contactEventService.findById(id);
-            Contact contact = contactEvent.getContact();
-            contact.setContactEvents(contact.getContactEvents().stream().filter(ce -> !ce.equals(contactEvent))
-                    .collect(Collectors.toSet()));
-            contactService.saveContact(contact);
-            contactEventService.deleteContactEvent(id);
-            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Contact event deleted");
-        } catch (NoSuchElementException e) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Contact event does not exist");
+            Optional<ContactEvent> optContactEvent = contactEventService.findById(id);
+            if (optContactEvent.isPresent()) {
+                ContactEvent contactEvent = optContactEvent.get();
+                Contact contact = contactEvent.getContact();
+                contact.setContactEvents(contact.getContactEvents().stream().filter(ce -> !ce.equals(contactEvent))
+                        .collect(Collectors.toSet()));
+                contactService.saveContact(contact);
+                contactEventService.deleteContactEvent(id);
+                return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Contact event deleted");
+            } else
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Contact event does not exist");
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Error");
